qemu: Do not expose the host model to a pseries guest
Libvirt should not expose host information to the guest by default so
this patch disables it.

However, if there are guest userspace tools that rely on this, an
administrator can still choose to allow the old behavior by setting
host_model='passthrough' in qemu.conf.

It is also possible to pass in an arbitrary string via the host_model
option, which can be useful in case libvirtd is running unprivileged.